Improve checksum management
When the checksum is not defined (hardcoded) the check would be
failing everytime the user modifies the number of grid points, or
request a size for which the checksum is not known.  Not properly
managing these cases would result with the application returning a
failure for bad checksum.

This commit resolves this case by propagating into Input a flag
indicating whether the number of grid points has been modified. If the
number of points has been modified (even if it stays the same as the
default for "small", "large", "XL" or "XXL"), then the checksum
comparison result is ignored, a warning is issued, but the application
returns a success return code.